requirements.txt
python-telegram-bot==21.6
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
duckduckgo-search>=7.2.1
python-dotenv==1.0.1
lxml[html_clean]==5.3.0
